[lang/ir] make __ink_ ident prefix check more general
